+++ frontend/techpick/src/app/(unsigned)/login/failed/page.tsx
@@ -1,15 +1,9 @@
-'use client';
-
-import { useEventLogger } from '@/hooks/useEventLogger';
 import Image from 'next/image';
-import Link from 'next/link';
+import { LoginButtonSection } from '../GoogleLoginButton';
 import {
   dividerStyle,
   failedDescriptionTextStyle,
-  googleLoginContainer,
-  kakaoLoginContainer,
   loginBlockContainer,
-  loginLink,
   pickBrandContainer,
   pickBrandContainerWithText,
   pickIconContainer,
@@ -17,14 +11,6 @@ import {
 } from '../page.css';
 
 export default function LoginFailedPage() {
-  const redirectUrl = encodeURIComponent(
-    process.env.NEXT_PUBLIC_REDIRECT_URL ?? '',
-  );
-
-  const { trackEvent: trackLoginButtonClick } = useEventLogger({
-    eventName: 'login_failed_page_login_button_click',
-  });
-
   return (
     <div className={screenContainer}>
       <div className={loginBlockContainer}>
@@ -38,50 +24,15 @@ export default function LoginFailedPage() {
                 objectFit={'contain'}
               />
             </div>
-            <h1>SIGN UP</h1>
+            <h1 style={{ fontSize: '40px', flexShrink: 0, flexGrow: 1 }}>
+              SIGN UP
+            </h1>
           </div>
         </div>
         <hr className={dividerStyle} />
-        <div style={{ paddingTop: '36px', paddingBottom: '8px' }}>
-          <div className={googleLoginContainer}>
-            <Link
-              className={loginLink}
-              href={`${process.env.NEXT_PUBLIC_API}/login/google?redirect_url=${redirectUrl}`}
-              onClick={() => {
-                trackLoginButtonClick();
-              }}
-            >
-              <Image
-                style={{ filter: 'brightness(100)' }}
-                src={'/image/logo_google.png'}
-                alt="Google Logo"
-                width={20}
-                height={20}
-              />
-              <span>Sign up with Google</span>
-            </Link>
-          </div>
-          <div className={kakaoLoginContainer}>
-            <Link
-              className={loginLink}
-              href={`${process.env.NEXT_PUBLIC_API}/login/kakao?redirect_url=${redirectUrl}`}
-              onClick={() => {
-                trackLoginButtonClick();
-              }}
-            >
-              <Image
-                style={{ filter: 'invert(100%)' }}
-                src={'/image/logo_kakao.svg'}
-                alt="Kakao Logo"
-                width={20}
-                height={20}
-              />
-              Sign up with Kakao
-            </Link>
-          </div>
-          <div className={failedDescriptionTextStyle}>
-            <p>죄송합니다. 로그인에 실패했습니다. </p>
-          </div>
+        <LoginButtonSection />
+        <div className={failedDescriptionTextStyle}>
+          <p>죄송합니다. 로그인에 실패했습니다. </p>
         </div>
       </div>
     </div>
